utility/Node: Add two comments and change child_index_t typedef to uint32_t

+++ utility/Node.hpp
@@ -16,8 +16,8 @@ class Node
 {
     public:
         /* Dependent typedefs */
-        typedef uint16_t path_index_t;
-        typedef uint16_t child_index_t;
+        typedef uint16_t path_index_t; // Depth
+        typedef uint32_t child_index_t; // Breadth
 
         /* Member functions/methods */
         bool leaf() const;
